# internal/openapi/spec/public.yaml
openapi: 3.0.3
info:
  title: metareel Public API
  version: "1.0.0"
  description: >
    Every /api/v1 endpoint needs an API key, sent as `X-API-Key: <key>` or
    `Authorization: Bearer <key>` (unless keys are turned off in the admin
    console). Keys are created and rotated in the admin console under
    Settings & API keys. Requests are rate limited per key; over the limit
    the API returns 429 with Retry-After. Missing or invalid keys get 401.
servers:
  - url: /
security:
  - ApiKeyHeader: []
  - BearerKey: []
paths:
...
  /api/v1/titles:
    get:
      operationId: listTitles
      summary: List titles (paged)
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv_show]
        - name: q
          in: query
          required: false
          schema:
            type: string
            description: Name or slug substring
        - name: missing
          in: query
          required: false
          schema:
            type: string
            enum: [tmdb, imdb, rt]
            description: Only titles without this ID
        - name: match_status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, matched, unmatched, manual]
            description: Only titles in this matching state (the admin review queue uses unmatched)
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [id_desc, id_asc, name_asc, name_desc, updated_desc, updated_asc, last_ranked_desc, last_ranked_asc, rankings_desc, rankings_asc]
            description: Default id_desc (newest first)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
      responses:
        "200":
          description: Titles page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitlesPageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/{id}:
    get:
      operationId: getTitle
      summary: Get one scraped title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/{id}/rankings:
    get:
      operationId: getTitleRankingsById
      summary: Ranking history of one scraped title (mapped to TMDB or not)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/{kind}/{id}/rankings:
    get:
      operationId: getTitleRankings
      summary: Ranking history of one title, looked up by TMDB ID
      description: >
        Every Top 10 chart appearance of the title, oldest first, plus a
        per-chart summary. All FlixPatrol titles mapped to the TMDB ID are
        included. Returns 404 if no scraped title is mapped to the TMDB ID.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: TMDB ID (numeric)
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No title mapped to this TMDB ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/{kind}/{id}/ratings:
    get:
      operationId: getTitleRatings
      summary: IMDb and Rotten Tomatoes ratings of a title, looked up by TMDB ID
      description: >
        Served from the database when refreshed within RATINGS_TTL (12h by
        default), otherwise refreshed from JustWatch, MDBList (when
        MDBLIST_API_KEY is set) and Rotten Tomatoes' own search index. Rotten
        Tomatoes scores come from RT directly when it has a match (by stored
        slug, MDBList's RT link, or a title + year search), else from MDBList /
        JustWatch. Every rating each provider returned is listed in sources.
        If a refresh fails, the previous ratings are returned with stale=true.
      parameters:
        - $ref: "#/components/parameters/TmdbKindParam"
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: TMDB ID (numeric)
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
            description: Refresh from the providers even if the stored ratings are fresh
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingsResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No provider knows this TMDB title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Every provider failed and nothing was stored yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/tmdb/rankings:
    get:
      operationId: getTitleRankingsBatch
      summary: Ranking history of several titles, looked up by TMDB ID
      description: >
        Batch version of getTitleRankings. Items come back in request order;
        IDs with no mapped title are listed in not_found.
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
            example: movie:425,tv:154385
            description: Comma-separated kind:tmdb_id pairs (kind is movie or tv), at most 50
        - $ref: "#/components/parameters/HistoryCountryParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
      responses:
        "200":
          description: Ranking histories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleRankingsBatchResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/titles/new/{country}/{service}:
    get:
      operationId: getNewTitles
      summary: Titles added to a streaming service, per day (live from JustWatch)
      description: >
        One JustWatch query per day from `from` to `to` (default: the last 7
        days, at most 31). Newest day first. Seasons are listed under their
        show. Nothing is stored.
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
            description: ISO country code (e.g. US)
        - name: service
          in: path
          required: true
          schema:
            type: string
            description: >
              FlixPatrol slug (netflix, hbo-max, disney, amazon-prime,
              paramount-plus, peacock, apple-tv; resolved to the country's
              JustWatch code) or any JustWatch package short / technical name
              from /api/v1/services/{country} (e.g. nfx, ppa)
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [movie, tv]
        - name: language
          in: query
          required: false
          schema:
            type: string
            default: en
            description: 2-letter language for titles, descriptions and genres
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
            description: Default today (UTC)
      responses:
        "200":
          description: New titles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleasesResponse"
        "400":
          description: Validation error (country, service, dates, type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: JustWatch request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/services/{country}:
    get:
      operationId: listServices
      summary: Streaming services JustWatch knows in a country
      parameters:
        - name: country
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicesResponse"
        "400":
          description: Validation error (country, service, dates, type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: JustWatch request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerKey:
      type: http
      scheme: bearer
      description: 'The same API key, as "Authorization: Bearer <key>"'
...
  parameters:
...
    TmdbKindParam:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [movie, tv]
        description: TMDB namespace of the ID
    HistoryCountryParam:
      name: country
      in: query
      required: false
      schema:
        type: string
        description: ISO country code (e.g. US); omit for all countries
    FromParam:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
        description: First chart date to include (YYYY-MM-DD)
    ToParam:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
        description: Last chart date to include (YYYY-MM-DD)
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required: [error]
...
    TitlesPageResponse:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TitleResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
      required: [items, total, limit, offset]
...